requirements.txt
pyautogui==0.9.54
keyboard==0.13.5
pyperclip==1.8.2
Pillow==10.0.0